Invoice Parser/requirements.txt
Flask
python-dotenv
celery
redis
google-generativeai
python-docx
PyMuPDF
Pillow
psycopg2-binary
pytz
